requirements.txt
asgiref==3.8.1
certifi==2024.8.30
chardet==3.0.4
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.2
